ホーム画面に追加ボタンをaboutからlinksページに移動
