+++ i18n/ja/about.js
@@ -12,13 +12,6 @@ export default {
         "アカウント登録やログイン不要で誰でも譜面を作成することができます。",
       content4:
         "作成した譜面はサーバーに保存され、SNS などで簡単に共有して他の人に遊んでもらうこともできます。",
-      installDesc:
-        "Android および iOS のタブレットやスマートフォンでは、" +
-        "Falling Nikochan をホーム画面に追加してアプリとして起動することにより、" +
-        "フルスクリーンでより快適に遊ぶことができます。",
-      install: "ホーム画面に追加",
-      installIOS:
-        "(iOS では Safari の共有メニューから「ホーム画面に追加」を選択してください。)",
     },
     2: {
       title: "遊び方",
